requirements.txt
# local package
-e .

# external requirements
boto3 == 1.34.69
pillow == 10.2.0
boto3 == 1.34.74
pandas == 2.2.1
requests == 2.31.0
scikit-learn == 1.4.1
matplotlib == 3.8.3
nltk == 3.8.1
spacy == 3.7.4
python-dotenv == 1.0.1
ipykernel == 6.29.3

